Use cumulated runtime to compute op/s
The current global time of the benchmark is currently used, which includes a
lot of the time in the client and the executor.

This instead only count for the cumulated time of all requests, and average it
on the number of workers being used. It obviously gives bigger number, but also
closer to the actual performance that you can expect from your Gnocchi API
server.