benchmark: generate just enough point for a batch
The current measures generation creates the number of points passed as --count
and split that into batch. This means the master process is so busy that it
cannot generates data fast enough for workers to consume.

Let's use a faster way by pre-generating a list of the batch size and use that
for all workers. It's good enough for testing the API response time.