Show and parse timestamps in local timezone by default
Many users are confused by the fact the returned timezone by the CLI is in UTC
where they expect to have local timezone. This fixes that by default while
keeping a --utc flag so it's possible to have UTC if wanted as output.

This patches handle all timestamps without specified timezone as being in local
timezone, which is the behavior of other database systems, such as PostgreSQL.

Sem-Ver: api-break