Do not send Accept on most delete API calls
fixes #124

since  Pecan 1.4.0 (and WebOb 1.8), the validation of Accept header
became more strict.
Most delete APIs in Gnocchi are not exposed as emitting JSON
(except for resource batch delete).
On the other hand, the base Manager in gnocchiclient by default
always sends "Accept: application/json" header.

As a result, currently corresponding 'delete' gnocchiclient commands are
failing with HTTP 406 Not Acceptable when executed agains Gnocchi that
is installed with fresh versions of Pecan >= 1.4.0 as a dependency.

This patch forces gnocchiclient to explicitly send "Accept: None" header
on those DELETE API calls that do not accept JSON:
- archive policy delete
- archive policy rule delete
- resource type delete
- resource delete
- metric delete