Adjust Metrics Not Found exception
Gnocchi may present one or the other response error for metrics not found depending of its version.

```
{"code": 404, "title": "Not Found", "description": {"cause": "Metrics not found", "detail": ["<metric name>"]}

```

```
{"code": 400, "description": {"cause": "Metrics not found", "detail": ["<metric name>"]}, "title": "Bad Request"}
```

Therefore, the exception handling in Gnocchi client is not correct, which causes the MetricsNotFound exception not to be presented when calling the Gnocchi Aggregates API. Instead a simple Not Found is presented to the user. This patch proposes a fix for that.

I maitained the previous workflow as well to handle situation that we have not seen yet with the description in the body as a String.