Drop implicit distutils dependency
Python 3.12 stopped shipping distutils:
https://docs.python.org/3/whatsnew/3.12.html
There're couple of places where gnocchiclient explicitly
imports from distutils - which now does not work.

Luckily, those imports were for a single reson only -
a strtobool function...
Let's just copy it over to utils.py and live with it,
not that much of a burden to support own copy.