Add Python 3.10 and 3.12 support
Fix python-gnocchiclient on Python 3.12 by replacing usage
of distutils.util.strtobool with a new function in
gnocchiclient.utils that does the same thing
(except returning an actual boolean instead of an integer).

Added Python 3.12 to the package Trove classifiers and tox.ini.

Also added Python 3.10 to the package Trove classifiers and tox.ini,
because it is generally an anti-pattern for a package to advertise
Python 3.8-3.9 and Python 3.11+ support without also supporting
Python 3.10.

Fix pep8 checks by suppressing the new A005 rule.